+++ packages/terra-date-input/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Changed
+  * Fix to prevent non-numeric characters from being entered in Safari.
+
 ## 1.30.0 - (July 21, 2021)
 
 * Changed